refactor: comment_evaluatorを評価基準ごとに分割
- 約1000行の単一ファイルを評価基準ごとの小さなモジュールに分割
- BaseEvaluatorクラスを導入して共通インターフェースを定義
- 各評価ロジック（関連性、創造性、自然さ、適切性、エンゲージメント、明確性、一貫性、独自性）を独立したモジュールに
- メインのcomment_evaluator.pyを260行に削減（約74%削減）
- 機能と出力は変更なし
